reverted changes

+++ filesize.c
@@ -1,8 +1,11 @@
-#include <sys/types.h>
-#include <sys/stat.h>
 #include "fsutil.h"
 
-long file_size(const char *file_name) {
+#include <sys/types.h>
+#include <sys/stat.h>
+
+
+int file_size(const char *file_name)
+{
   struct stat fileStat;
   int err = stat( file_name, &fileStat );
   if (0 != err) return 0;
